refactor(test): Reorganize tests
