+++ tests/unit/test_workflows.py
@@ -28,36 +28,36 @@ from tracecat.auth import (
     authenticate_user,
     authenticate_user_or_service,
 )
-from tracecat.db import TRACECAT__DB_URI
 from tracecat.runner.app import app as runner_app
 from tracecat.runner.app import valid_workflow
 
 TEST_WORKFLOW_RUN_TIMEOUT = 60  # seconds
 TEST_USER_ID = "3f1606c4-351e-41df-acb4-fb6e243fd071"
 
-
-client = TestClient(app=app)
-
-
-# Override authentication dependencies
-app.dependency_overrides[authenticate_user] = lambda: Role(
-    type="user", user_id=TEST_USER_ID
-)
-app.dependency_overrides[authenticate_service] = lambda: Role(
-    type="service", service_id="tracecat-runner"
-)
-app.dependency_overrides[authenticate_user_or_service] = lambda: Role(
-    type="user", user_id=TEST_USER_ID, service_id="tracecat-runner"
-)
-runner_app.dependency_overrides[authenticate_service] = lambda: Role(
-    type="service", user_id="tracecat-api"
-)
-runner_app.dependency_overrides[valid_workflow] = lambda workflow_id: workflow_id
+client: TestClient
 
 
 # Define a fixture for the Runner server
-@pytest.fixture(scope="session", autouse=True)
+@pytest.fixture(scope="session")
 def start_servers():
+    global client
+    client = TestClient(app=app)
+
+    # Override authentication dependencies
+    app.dependency_overrides[authenticate_user] = lambda: Role(
+        type="user", user_id=TEST_USER_ID
+    )
+    app.dependency_overrides[authenticate_service] = lambda: Role(
+        type="service", service_id="tracecat-runner"
+    )
+    app.dependency_overrides[authenticate_user_or_service] = lambda: Role(
+        type="user", user_id=TEST_USER_ID, service_id="tracecat-runner"
+    )
+    runner_app.dependency_overrides[authenticate_service] = lambda: Role(
+        type="service", user_id="tracecat-api"
+    )
+    runner_app.dependency_overrides[valid_workflow] = lambda workflow_id: workflow_id
+
     thread = threading.Thread(
         target=lambda: uvicorn.run(app, host="localhost", port=8000, log_level="info")
     )
@@ -86,17 +86,17 @@ def create_test_db():
         # Insert data into database
         actions.write_database(
             table_name="action",
-            connection=TRACECAT__DB_URI,
+            connection=os.environ["TRACECAT__DB_URI"],
             if_table_exists="append",
         )
         webhooks.write_database(
             table_name="webhook",
-            connection=TRACECAT__DB_URI,
+            connection=os.environ["TRACECAT__DB_URI"],
             if_table_exists="append",
         )
         workflows.write_database(
             table_name="workflow",
-            connection=TRACECAT__DB_URI,
+            connection=os.environ["TRACECAT__DB_URI"],
             if_table_exists="append",
         )
 
@@ -120,7 +120,7 @@ def pytest_generate_tests(metafunc):
     # Directly use the fixture data for parametrization
     query = "SELECT workflow_id, action_id FROM webhook"
     triggers = (
-        pl.read_database_uri(query, uri=TRACECAT__DB_URI, engine="adbc")
+        pl.read_database_uri(query, uri=os.environ["TRACECAT__DB_URI"], engine="adbc")
         .join(pl.read_ndjson("tests/data/trigger_payloads.ndjson"), on="action_id")
         .to_dicts()
     )
